backend/requirements.txt
flask==3.1.3
flask-cors==6.0.2
python-dotenv==1.2.1
pymysql==1.1.2
bcrypt==5.0.0
PyJWT==2.11.0
